% Generated by build script
\name{imf_indicators}
\alias{imf_indicators}
\title{List available IMF indicators}
\usage{
imf_indicators()
}
\arguments{
\item{query}{Optional search term to filter by name}
}
\value{tibble: id, label, description, unit, dataset}
\description{List available IMF indicators}
